+++ Fluent.Ribbon/ILargeIconProvider.cs
@@ -1,0 +1,13 @@
+﻿namespace Fluent
+{
+    /// <summary>
+    /// Inferface for controls which provide a large icon.
+    /// </summary>
+    public interface ILargeIconProvider
+    {
+        /// <summary>
+        /// Gets or sets the large icon.
+        /// </summary>
+        object LargeIcon { get; set; }
+    }
+}